requirements.txt
Flask
ultralytics
opencv-python-headless
Pillow
